feat: setup workspace at build time
